+++ src/main/java/io/hurem/domain/employees/EmployeesRepository.java
@@ -2,6 +2,6 @@ package io.hurem.domain.employees;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeesRepository extends JpaRepository<Employees, Long>  {
+public interface EmployeesRepository extends JpaRepository<Employees, Integer>  {
 
 }
