+++ src/main/java/io/hurem/domain/jobs/JobsRepository.java
@@ -2,6 +2,6 @@ package io.hurem.domain.jobs;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobsRepository extends JpaRepository<Jobs, Long>  {
+public interface JobsRepository extends JpaRepository<Jobs, Integer>  {
 
 }
